refactor: mapping into seperate file
